requirements.txt
Flask==3.0.1
Flask-Dance==7.1.0
python-dotenv==1.0.1
requests-oauthlib==2.0.0
pinecone-client==3.2.2
gunicorn==21.2.0

